+++ fnv1/hash_test.go
@@ -5,13 +5,14 @@ import (
 	"testing"
 
 	"github.com/segmentio/fasthash"
+	"github.com/segmentio/fasthash/fasthashtest"
 )
 
 func TestHash64(t *testing.T) {
-	fasthash.TestHashString64(t, "fnv1", fasthash.HashString64(fnv.New64), HashString64)
-	fasthash.TestHashUint64(t, "fnv1", fasthash.HashUint64(fnv.New64), HashUint64)
+	fasthashtest.TestHashString64(t, "fnv1", fasthash.HashString64(fnv.New64), HashString64)
+	fasthashtest.TestHashUint64(t, "fnv1", fasthash.HashUint64(fnv.New64), HashUint64)
 }
 
 func BenchmarkHash64(b *testing.B) {
-	fasthash.BenchmarkHashString64(b, "fnv1", fasthash.HashString64(fnv.New64a), HashString64)
+	fasthashtest.BenchmarkHashString64(b, "fnv1", fasthash.HashString64(fnv.New64a), HashString64)
 }
