+++ fnv1a/hash32_test.go
@@ -5,13 +5,14 @@ import (
 	"testing"
 
 	"github.com/segmentio/fasthash"
+	"github.com/segmentio/fasthash/fasthashtest"
 )
 
 func TestHash32(t *testing.T) {
-	fasthash.TestHashString32(t, "fnv1a", fasthash.HashString32(fnv.New32a), HashString32)
-	fasthash.TestHashUint32(t, "fnv1a", fasthash.HashUint32(fnv.New32a), HashUint32)
+	fasthashtest.TestHashString32(t, "fnv1a", fasthash.HashString32(fnv.New32a), HashString32)
+	fasthashtest.TestHashUint32(t, "fnv1a", fasthash.HashUint32(fnv.New32a), HashUint32)
 }
 
 func BenchmarkHash32(b *testing.B) {
-	fasthash.BenchmarkHashString32(b, "fnv1a", fasthash.HashString32(fnv.New32a), HashString32)
+	fasthashtest.BenchmarkHashString32(b, "fnv1a", fasthash.HashString32(fnv.New32a), HashString32)
 }
